+++ sdk/ai/Azure.AI.Projects.Agents/CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fixed the `StopSession` and `StopSessionAsync` calls.
 
 ### Other Changes
+- Updated the `OpenAI` package dependency to `2.11.0`.
 
 ### Sample Updates
 - Added sample demonstrating disabling and enabling Hosted Agent.